+++ src/Components/TodoList/TodoList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { TodoInfo } from '../TodoInfo/TodoInfo';
+import { TodoItem } from '../TodoItem/TodoItem';
 import { Todo } from '../../Types/Todo';
 
 type Props = {
@@ -11,7 +11,7 @@ export const TodoList: React.FC<Props> = ({ todos }) => {
   return (
     <ul className="todo-list" data-cy="todoList">
       {todos.map((todo: Todo) => (
-        <TodoInfo todo={todo} key={todo.id} />
+        <TodoItem todo={todo} key={todo.id} />
       ))}
     </ul>
   );
